Working tests for logging disabled
